rag_service/requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
pydantic
cohere
requests
